Rename typoed file
