+++ src/UI/EmailSignupHTMLTooltip.js
